+++ src/constants.ts
@@ -11,6 +11,11 @@ export const QW_LOG = vscode.window.createOutputChannel("Quarto Wizard", { log: 
 export const QW_RECENTLY_INSTALLED = "recentlyInstalledExtensions";
 
 /**
+ * Key for storing recently used templates.
+ */
+export const QW_RECENTLY_USED = "recentlyUsedTemplates";
+
+/**
  * URL to the Quarto extensions CSV file.
  */
 export const QW_EXTENSIONS =
@@ -24,7 +29,7 @@ export const QW_EXTENSIONS_CACHE = "quarto_wizard_extensions";
 /**
  * Cache duration for the Quarto extensions JSON (default to 1 hour).
  */
-export const QW_EXTENSIONS_CACHE_TIME = 1 * 60 * 60 * 1000;
+export const QW_EXTENSIONS_CACHE_TIME = 0 * 60 * 60 * 1000;
 
 /**
  * Markdown Lint extension identifier.